Crate rl_gl

is OpenGL Wrapper that attempts to be Type Safe as well as preferment

Reexports

pub use self::attributes::Attribute;

pub use self::attributes::AttributeKind;

pub use self::attributes::DescribeAttributes;

Modules

attributes
Tools to communicate how Rust structs and types can be used by the OpenGL driver with Vertices

buffer
Handles memory allocation for the GPU as well as transferring the Description of of the data, so OpenGL can communicate to it.

errors
Errors from interactive with OpenGL Driver

program
Compiled Shader Pipeline

raw
Unsafe OpenGL calls, enums, and types

Traits

GlObject
All OpenGL objects have an id which uses to tell the driver to perform commands on them. However, sometimes questions needed to be asked about an object after original abstraction has been Dropped