Trait rl_gl::attributes::DescribeAttributes

pub trait DescribeAttributes: Clone {
    unsafe fn attributes() -> Vec<Attribute>
    where
        Self: Sized;
}

creates a series of OpenGL friendly attributes use to bind vertices to buffers

Example Manual Implementation

use rl_gl::DescribeAttributes;
use rl_gl::attributes::{Attribute, AttributeSize, AttributeKind};

#[derive(Clone, Debug)]
struct Vertices {
    position: [f32; 3],
    color: [u8; 4],
}

impl DescribeAttributes for Vertices {
    unsafe fn attributes() -> Vec<Attribute> {
        use std::mem;
        vec![
            Attribute::new(
                AttributeSize::Three,
                AttributeKind::Float,
                false,
                mem::size_of::<Vertices>(),
                &(*(std::ptr::null() as *const Vertices)).position as *const _ as usize
            ),
            Attribute::new(
                AttributeSize::Four,
                AttributeKind::UnsignedByte,
                false,
                mem::size_of::<Vertices>(),
                &(*(std::ptr::null() as *const Vertices)).color as *const _ as usize
            ),
        ]
    }
}

However, you do not need to implement DescribeAttributes for each struct you want to turn into vertices.

Example Derive

use std::mem;
use rl_gl::DescribeAttributes;
use rl_gl::attributes::{ Attribute, AttributeKind, AttributeSize };

#[derive(Clone, Debug, DescribeAttributes)]
struct Vertices {
    position: [f32; 3],
    color: [u8; 4],
}

unsafe {
  let attributes = Vertices::attributes();

  assert_eq!(attributes.len(), 2);
  assert_eq!(
      attributes[0],
      Attribute::new(
          AttributeSize::Three,
          AttributeKind::Float,
          false,
          mem::size_of::<Vertices>(),
          &(*(std::ptr::null() as *const Vertices)).position as *const _ as usize,
      )
  )
}

Required Methods

unsafe fn attributes() -> Vec<Attribute> where
Self: Sized,

gets an Vec of Attributes describing each field. the Vec should be ordered to the order of the fields when implemented