Module rl_gl::buffer

Handles memory allocation for the GPU as well as transferring the Description of of the data, so OpenGL can communicate to it.

Structs

BoundBufferObject
Is used to scope a OpenGL BindBuffer. It unbinds the buffer when it goes out of scope.

BoundVertexArrayObject
Is used to scope a OpenGL BindVertexArrays. It unbinds the buffer when it goes out of scope.

BufferConfiguration
Allows for Configuration and Building Buffer Objects.

BufferConfigurationWithIndex
Allows for Configuration and Building Buffer Objects with Indices

BufferObject
Reference to stored unformatted data on GPU.

BufferReference
Reference to OpenGL buffer.

VertexArrayObject
Reference to Vertex description of data stored inside of Buffer Object

Enums

BufferKind
Used to specify the kind of buffer a BufferObject is bound too.

Traits

BufferBuilder
Buffers Reference can be build from objects that implement this