Struct rl_gl::buffer::BufferObject

pub struct BufferObject { /* fields omitted */ }

Reference to stored unformatted data on GPU.

more from OpenGL API

Methods

impl BufferObject

pub fn new(kind: BufferKind) -> BufferObject

Generates new Buffer Object

more from the OpenGL API

pub fn bind<'a>(&'a mut self) -> BoundBufferObject<'a>

Temporary Binds the BufferObject to the OpenGL context. Although it isn't required, if you have a Vertex Array Buffer bounded you should pass it through the BufferObject bounded, so that the binding and unbinding behavior becomes deterministic

pub fn kind(&self) -> BufferKind

gets kind of Buffer Object