Struct rl_gl::program::shaders::CompiledShader

pub struct CompiledShader { /* fields omitted */ }

CompiledShader is an abstraction representation of a compiled GLSL shader

Notes

CompiledShader will tell the GPU to destroy itself whenever it is dropped

Methods

impl CompiledShader

pub fn new<T: Into<Vec<u8>>>(
src: T,
kind: ShaderKind
) -> GlResult<CompiledShader>

Creates and Compiles Shader.

pub fn is_valid(&self) -> GlResult<()>

Checks if the CompiledShader is an shader, has been successfully compiled, and isn't marked for deletion

pub fn kind(&self) -> ShaderKind

Returns the kind of shader the Abstraction expects the shader to be