Struct rl_gl::program::ShaderProgram

pub struct ShaderProgram(_);

ShaderProgram is an abstract representation of GLSL Object

Methods

impl ShaderProgram

pub fn new(
vertex_shader: &CompiledShader,
fragment_shader: &CompiledShader
) -> GlResult<ShaderProgram>

Creates a Shader Program by linking the given Vertex Shader and the Fragment Shader

pub fn set_to_current(&self)

Sets the program to the current program in the context