Trait rl_gl::GlObject

pub trait GlObject {
    fn as_gl_id(&self) -> GLuint;
}

All OpenGL objects have an id which uses to tell the driver to perform commands on them. However, sometimes questions needed to be asked about an object after original abstraction has been Dropped

Required Methods

fn as_gl_id(&self) -> GLuint

Gets the Id of the GlObejct